host: new configure options: --with-preferred-host-{ar|as|nm|ranlib}
With these configure options, it is possible to name the exact tools
to be chosen by slibtool when carrying out a native build task and given
that the tool variant had not been otherwise specified. This could come
in handy, for instance, on systems where ar(1) is installed by is not
the preferred native archiving tool.